requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
azure-cosmos>=4.5.0
gunicorn>=21.0.0
